requirements_ews.txt
# Enhanced Requirements for Early Warning System (EWS)
# Student Dropout Prevention System with ML Integration

# Core Flask Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
Flask-Talisman==1.0.0
Flask-Migrate==4.0.5

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Machine Learning & Data Science
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
joblib==1.4.0

# Background Tasks & Queue
celery==5.3.4
redis==5.0.1

# Web Framework & Frontend
streamlit==1.36.0
Werkzeug==3.0.1
Jinja2==3.1.2

# Forms & Validation
WTForms==3.1.1
email-validator==2.1.0

# Security & Authentication
bcrypt==4.1.2
PyJWT==2.8.0
cryptography==41.0.8

# Data Processing & Export
openpyxl==3.1.2
xlsxwriter==3.1.9
reportlab==4.0.7
Pillow==10.1.0

# HTTP & API
requests==2.31.0
urllib3==2.0.7
python-dotenv==1.0.0

# Logging & Monitoring
sentry-sdk[flask]==1.40.6

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Utilities
click==8.1.7
python-magic==0.4.27
validators==0.22.0

# Development & Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Production Server
gunicorn==21.2.0
eventlet==0.33.3

# Additional Utilities
marshmallow==3.20.2
marshmallow-sqlalchemy==0.29.0
python-socketio==5.10.0
schedule==1.2.0

# SMS Integration (Optional)
twilio==8.10.3

# Email Templates
markdown==3.5.1

# File Upload
python-multipart==0.0.6

# Rate Limiting
limits==3.13.0

# Caching
Flask-Caching==2.1.0

# API Documentation
flask-restx==1.3.0
flask-cors==4.0.0

# Data Validation
cerberus==1.3.4

# Background Task Scheduling
APScheduler==3.10.4

# Database Migrations
alembic==1.13.1

# Environment Configuration
python-decouple==3.8

# Image Processing (for student photos)
Pillow==10.1.0

# PDF Generation
fpdf2==2.7.8

# Excel Export
xlsxwriter==3.1.9

# CSV Processing
chardet==5.2.0

# JSON Schema Validation
jsonschema==4.19.2

# Timezone Handling
pytz==2023.3

# HTTP Client for external APIs
httpx==0.25.2

# WebSocket Support (for real-time notifications)
flask-socketio==5.3.6

# File Storage (AWS S3)
boto3==1.34.0

# Password Strength Checking
zxcvbn==4.4.2

# Session Management
Flask-Session==0.5.0
